backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
openai==1.68.2
pydantic==2.11.1
logfire
python-dotenv==1.1.0
httpx==0.28.1
